api/python/requirements.txt
Werkzeug==0.16.1
Flask==1.1.2
gunicorn==20.0.4
flask-restplus==0.13.0
google-cloud-datastore==2.0.1
google-cloud-kms==2.2.0
cryptography==3.2.1
crcmod==1.7
python-ldap==3.3.1
ldap3==2.8.1

